Измненение типов

+++ app/db/models/machine.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, BigInteger, Boolean
+from sqlalchemy import String, Integer, Boolean
 from sqlalchemy.orm import Mapped, mapped_column
 from datetime import datetime
 from app.db.base import Base
@@ -8,10 +8,10 @@ class machines(Base):
     __tablename__ = "machines"
 
     # 1. ID с автоинкрементом
-    id: Mapped[BigInteger] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
         
     type_machine: Mapped[String] = mapped_column(String, nullable=False)
-    number_machine: Mapped[BigInteger] = mapped_column(BigInteger, nullable=False)
+    number_machine: Mapped[int] = mapped_column(Integer, nullable=False)
     status: Mapped[Boolean] = mapped_column(Boolean, default=False, nullable=False)
