+++ app/db/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, BigInteger
+from sqlalchemy import String, BigInteger, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base
 
@@ -6,12 +6,12 @@ class users(Base):
     __tablename__ = "users"
 
     #1. ID primary_key с автоинкрементом
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+
 
     inidroom: Mapped[int] = mapped_column(BigInteger, nullable=False)
     idcards: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    
+
     # 3. Данные телеграм
     tg_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False)
     last_name: Mapped[str] = mapped_column(String, nullable=False)
